update android 15 sdk source

+++ com/android/org/conscrypt/metrics/Protocol.java
@@ -35,7 +35,7 @@ public enum Protocol {
     TLS_PROTO_FAILED(0xFFFF),
     ;
 
-    final byte id;
+    final int id;
 
     public int getId() {
         return this.id;
@@ -61,6 +61,6 @@ public enum Protocol {
     }
 
     private Protocol(int id) {
-        this.id = (byte) id;
+        this.id = id;
     }
 }
